requirements.txt
beautifulsoup4==4.12.2
matplotlib==3.5.2
numpy==1.25.2
pandas==1.5.3
Pillow==10.0.0
plotly==5.9.0
PyYAML==6.0.1
requests==2.28.1
streamlit==1.12.2
streamlit_authenticator==0.2.1
yfinance
